feat: 북마크 카드 세로 길이 유동적으로 변하는 레이아웃 구현
